Change 'run' defaults for dialect and input
Dynamically set the default for dialect and input based on the DBMS.
The user can still override the defaults.

This also requires DSHOME to be set.  Added --tpcdstools long option to
set DSHOME.  DSHOME can still be set in the environment.

Removed the unused dialect and input related variables from get-config.