Add/replace short options with long options
Hopefully makes things easier to use.

Also removed unused seed options for various load scripts.
Specification is clear when to set a seed for the performance tests so
just use a random seed for the load test when using the simple 'run'
script.  A seed can be specified if the load test is run directly.