generate-queries: add --split flag to generate query per file
This by itself isn't to spec since it doesn't retain the correct query
order.  Extra work needs to be done if that is a requirement and you
need the queries split into individual files.

Nevertheless this may be helpful for testing purposes where a query per
file more helpful, or being able to select individual generated queries
by file as opposed to editing a single file with all the queries.

Note that the dsqgen has a side effect where a RNDSEED of 0 generates
the same queries for all streams when more than 1 is used.